[+BUGFIX] FLOW3 Distribution: Ship PackageStates.yaml
The fix for #3989 was to generate PackageStates.yaml and in August 2010
we had removed that file from the distribution.

With the dependency on PHPUnit being installed for the Testing package
this breaks if PHPUnit is not available, as we don't have dependency
checking. Thus add the file again and deactivate Testing in it.

Change-Id: Icf22efc1d0c786a1f052e1829676c40180b4c8c1
Fixes: #11207