[TASK] Removed minimum stability setting
This removes the minimum-stability so tagged releases of dependencies
are used. It explicitly allows dev versions for typo3/fluid, typo3/party
(because we want to use them as dev-master) and doctrine/migrations
(because no suitable stable release exists).

This change speeds up installation (because zip archives will be used
instead of git clones) and avoids surprises due to possibly unstable
3rd party code being pulled in.

Change-Id: Ic6d0ab06b8013718a34efa7a93689714b1016c55
Related: #47808
Releases: master, 2.0