[BUGFIX] Pin doctrine/migrations to known working version
With doctrine/migrations not having had a stable branch for years,
Flow always required dev-master. Now that conflicts with the version
of doctrine/orm used with Flow 2.x, so we need to pin the required
version to the last known working state.

Unfortunately this needs to be repeated in the root manifest.

Change-Id: I97c58796fc983b850719905aaf57317a209be97b
Releases: 2.3, 2.2, 2.1
Related: FLOW-300