[BUGFIX] Pin doctrine/migrations to tagged working version
With doctrine/migrations not having had a stable branch for years,
Flow always required dev-master. We pinned it to a version known as
the last known working state.

But due to the way composer works, only tagged version results in the
expected behavior, see https://github.com/composer/composer/issues/4082.

Unfortunately this needs to be repeated in the root manifest.

Change-Id: Ib1d4c22eb18ef4e74a32ea8e53520bc0db68afcc
Releases: 2.3, 2.2, 2.1
Related: FLOW-300